+++ src/apis/vector.ts
@@ -184,6 +184,7 @@ export default class VectorAPI implements VectorStorage {
 					undefined,
 					undefined,
 					undefined,
+					undefined,
 					'vector'
 				);
 				if (resp.status === 404) {